Fixed TrcTrackSet::getTrackSetNumber().
The getter used findAndGetUint16() although Track Set Number
(0066,0105) has VR UL, so it always failed with EC_IllegalCall and
never returned a value. Changed the parameter and the lookup to
Uint32. Added a regression check to the dcmtract_create test.

Thanks to GitHub user "ppmartel" (Phil Martel) for the report and fix.

+++ dcmtract/include/dcmtk/dcmtract/trctrackset.h
@@ -1,6 +1,6 @@
 /*
  *
- *  Copyright (C) 2016-2025, Open Connections GmbH
+ *  Copyright (C) 2016-2026, Open Connections GmbH
  *  All rights reserved.  See COPYRIGHT file for details.
  *
  *  This software and supporting documentation are maintained by
@@ -123,7 +123,7 @@ public:
     *  @param  pos Index of the value to get (0..vm-1)
     *  @return EC_Normal if successful, an error code otherwise
     */
-  virtual OFCondition getTrackSetNumber(Uint16 &value,
+  virtual OFCondition getTrackSetNumber(Uint32 &value,
                                         const unsigned long pos = 0) const;
 
   /** Get Track Set Label
